Create README.md
This project is a Document Upload and Retrieval System designed to simplify the submission, storage, and access of important documents such as journals, certificates, and ID proofs. It allows users to upload any document or image file, which is then stored securely in a server folder and associated with a system-generated Unique ID.

The key benefit of this system is that users do not need to remember file names or file paths. Instead, they just save the Unique ID displayed after uploading. Later, they can use this same ID to retrieve or verify their document, making the entire process faster, easier, and error-free.